@@ -1,6 +1,6 @@
 package edu.upc.essi.dtim.odin.models.rest;
 
-import edu.upc.essi.dtim.odin.models.mongo.DataSource;
+import edu.upc.essi.dtim.odin.bootstrapping.DataSource;
 import edu.upc.essi.dtim.nextiadi.models.Alignment;
 import lombok.Getter;
 import lombok.Setter;
