@@ -1,15 +1,15 @@
 package edu.upc.essi.dtim.odin.services.impl;
 
 import edu.upc.essi.dtim.odin.repository.DataSourcesRepository;
-import edu.upc.essi.dtim.odin.utils.jena.GraphOperations;
+//import edu.upc.essi.dtim.odin.utils.jena.GraphOperations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
 public class OMQService {
 
-    @Autowired
-    private GraphOperations graphOperations;
+//    @Autowired
+//    private GraphOperations graphOperations;
     @Autowired
     private DataSourcesRepository dsRepository;
 
