@@ -1,15 +1,16 @@
 package edu.upc.essi.dtim.odin.services.impl;
 
+import edu.upc.essi.dtim.Graph;
 import edu.upc.essi.dtim.odin.config.DataSourceTypes;
 import edu.upc.essi.dtim.odin.config.vocabulary.SourceGraph;
 import edu.upc.essi.dtim.odin.models.mongo.helpers.Attribute;
-import edu.upc.essi.dtim.odin.models.mongo.DataSource;
+import edu.upc.essi.dtim.odin.bootstrapping.DataSource;
 import edu.upc.essi.dtim.odin.models.mongo.Wrapper;
 import edu.upc.essi.dtim.odin.repository.DataSourcesRepository;
 import edu.upc.essi.dtim.odin.repository.WrapperRepository;
 import edu.upc.essi.dtim.odin.services.omq.WrapperI;
-import edu.upc.essi.dtim.odin.utils.jena.GraphOperations;
-import org.apache.jena.rdf.model.Model;
+//import edu.upc.essi.dtim.odin.utils.jena.GraphOperations;
+import edu.upc.essi.dtim.odin.storage.JenaConnection;
 import org.apache.jena.vocabulary.RDF;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -24,8 +25,10 @@ import java.util.stream.Collectors;
 public class WrapperService {
 
 
+//    @Autowired
+//    private GraphOperations graphOperations;
     @Autowired
-    private GraphOperations graphOperations;
+    private JenaConnection graph;
     @Autowired
     private DataSourcesRepository dataSourcesRepository;
     @Autowired
@@ -53,32 +56,32 @@ public class WrapperService {
     (triple <http://www.essi.upc.edu/~snadal/BDIOntology/Source/DataSource/Teams/teamName> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://www.essi.upc.edu/~snadal/BDIOntology/Source/Attribute>)
     (triple <http://www.essi.upc.edu/~snadal/BDIOntology/Source/DataSource/Teams/id> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://www.essi.upc.edu/~snadal/BDIOntology/Source/Attribute>)
     * */
-    public Model create(String wname, Attribute[] attrNames, String dataSourceId) {
+    public Graph create(String wname, Attribute[] attrNames, String dataSourceId) {
         Optional<DataSource> optionalDataSources = dataSourcesRepository.findById(dataSourceId);
         if (optionalDataSources.isPresent()) {
             DataSource ds = optionalDataSources.get();
             String wIRI = createWrapperIri(wname);
-            graphOperations.addTriple(ds.getIri(), wIRI, RDF.getURI() + "type", SourceGraph.WRAPPER.val());
-            graphOperations.addTriple(ds.getIri(), ds.getIri(), SourceGraph.HAS_WRAPPER.val(), wIRI);
+            graph.persistent().addTriple(ds.getIri(), wIRI, RDF.getURI() + "type", SourceGraph.WRAPPER.val());
+            graph.persistent().addTriple(ds.getIri(), ds.getIri(), SourceGraph.HAS_WRAPPER.val(), wIRI);
             for (Attribute attribute: attrNames) {
                 String attName = attribute.getName();
                 String attIRI = ds.getIri() + '/' + attName;
-                graphOperations.addTriple(ds.getIri(), attIRI, RDF.getURI() + "type", SourceGraph.ATTRIBUTE.val());
-                graphOperations.addTriple(ds.getIri(), wIRI, SourceGraph.HAS_ATTRIBUTE.val(), attIRI);
+                graph.persistent().addTriple(ds.getIri(), attIRI, RDF.getURI() + "type", SourceGraph.ATTRIBUTE.val());
+                graph.persistent().addTriple(ds.getIri(), wIRI, SourceGraph.HAS_ATTRIBUTE.val(), attIRI);
             }
 
-            return graphOperations.getGraph(ds.getIri());
+            return graph.persistent().getGraph(ds.getIri());
         }
         return null;
     }
 
-    public Wrapper createWrapperBootstrapped(Model bootstrap, String dsName, String dsID) {
+    public Wrapper createWrapperBootstrapped(Graph bootstrap, String dsName, String dsID) {
 
         // TODO: probably wrapper name should be given by user or generated incrementally for x wrappers (future)
         String wIRI = createWrapperIri(dsName +"Wrapper");
 
-        graphOperations.addModel(wIRI, bootstrap);
-        graphOperations.addTriple(wIRI,wIRI, RDF.type.getURI(), SourceGraph.WRAPPER.val());
+        graph.persistent().addModel(wIRI, bootstrap);
+        graph.persistent().addTriple(wIRI,wIRI, RDF.type.getURI(), SourceGraph.WRAPPER.val());
 
         Wrapper w = new Wrapper();
         w.setName(dsName+"Wrapper");
@@ -112,7 +115,7 @@ public class WrapperService {
         if (optionalWrapper.isPresent()) {
             Wrapper w = optionalWrapper.get();
             //Delete from Jena
-            graphOperations.removeGraph(w.getIri());
+            graph.persistent().deleteGraph(w.getIri());
             //Delete all lavMappings with lavMappingId
 //            lavMappingRepository.deleteById(w.getLavMappingId());
         }
@@ -124,14 +127,14 @@ public class WrapperService {
         //Delete from Jena
         //DsIri = http://www.essi.upc.edu/~snadal/BDIOntology/Source/DataSource/asd
         //WIri = http://www.essi.upc.edu/~snadal/BDIOntology/Source/Wrapper/w1
-        graphOperations.deleteTriplesWithSubject(ds.getIri(), createWrapperIri(w.getName()));
-        graphOperations.deleteTriplesWithObject(ds.getIri(), createWrapperIri(w.getName()));
+        graph.persistent().deleteTriplesWithSubject(ds.getIri(), createWrapperIri(w.getName()));
+        graph.persistent().deleteTriplesWithObject(ds.getIri(), createWrapperIri(w.getName()));
         //Eliminación de attributos
         for (Attribute a : w.getAttributes()) {
             String attName = a.getName();
             String attIRI = ds.getIri() + '/' + attName;
-            graphOperations.deleteTriplesWithSubject(ds.getIri(), attIRI);
-            graphOperations.deleteTriplesWithObject(ds.getIri(), attIRI);
+            graph.persistent().deleteTriplesWithSubject(ds.getIri(), attIRI);
+            graph.persistent().deleteTriplesWithObject(ds.getIri(), attIRI);
         }
 
 
