@@ -1,7 +1,7 @@
 package edu.upc.essi.dtim.odin.services.omq;
 
 import edu.upc.essi.dtim.odin.config.DataSourceTypes;
-import edu.upc.essi.dtim.odin.models.mongo.DataSource;
+import edu.upc.essi.dtim.odin.bootstrapping.DataSource;
 
 import java.util.List;
 
