@@ -1,36 +1,49 @@
-package edu.upc.essi.dtim.odin.config.db;
+package edu.upc.essi.dtim.odin.storage;
+import edu.upc.essi.dtim.odin.storage.filestorage.FileSystemStorageService;
+import edu.upc.essi.dtim.odin.storage.filestorage.StorageProperties;
+import edu.upc.essi.dtim.odin.storage.graph.GraphStore;
 import org.apache.jena.query.Dataset;
 import org.apache.jena.tdb.TDBFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.DependsOn;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PreDestroy;
+import javax.xml.crypto.Data;
 
 @Component
+@DependsOn("fileSystemStorageService")
 public class JenaConnection {
 
-    @Value("${dataStorage.landingZone.dir}")
-    private String dir;
-    @Value("${dataStorage.landingZone.jena.name}")
-    private String name;
+//    @Value("${dataStorage.landingZone.dir}")
+//    private String dir;
+//    @Value("${dataStorage.landingZone.jena.name}")
+//    private String name;
 
-    private static JenaConnection instance = new JenaConnection();
+//    private static JenaConnection instance = new JenaConnection();
     private static final Logger LOGGER = LoggerFactory.getLogger(JenaConnection.class);
 
-    private Dataset persistentDataset;
-    private Dataset temporalDataset;
+    private GraphStore persistentDataset;
+    private GraphStore temporalDataset;
 
-    public Dataset getTDBDataset() {
-        LOGGER.info("getTDBDataset()");
-        LOGGER.info("Jena Dir: {}" , dir);
-        LOGGER.info("Jena Name: {}" , name);
-        if (persistentDataset == null) {
+    Dataset tmp;
+
+    public Dataset getTmp(){return tmp;}
+
+    @Autowired
+    public JenaConnection(StorageProperties properties, FileSystemStorageService io) {
+        io.init();
+        System.out.println("jena const");
+        open(properties.getPersistentDir()+"/"+properties.getJena(),properties.getTemporalDir()+"/"+properties.getJena());
+
+
+        if (tmp == null) {
             try {
-                persistentDataset = TDBFactory.createDataset(dir+"/"+name);
+                tmp = TDBFactory.createDataset(properties.getTemporalDir()+"/tmp");
 
-                if(persistentDataset== null){
+                if(tmp== null){
                     LOGGER.info("DATASET NULO");
                 }
 
@@ -40,8 +53,66 @@ public class JenaConnection {
 
 
         }
+
+
+        System.out.println("end jena const");
+    }
+
+//    public void init(){
+//        System.out.println("jena init");
+//    }
+
+    public void open(String persistent, String temporal){
+        LOGGER.info("getTDBDataset()");
+        LOGGER.info("Jena Dir: {}" , persistent);
+        LOGGER.info("Jena temporal: {}" , temporal);
+
+        try {
+            persistentDataset = GraphStore.createGraphStore(persistent);
+//            persistentDataset = TDBFactory.createDataset(persistent);
+            if(persistentDataset== null){
+                LOGGER.info("DATASET NULO");
+            }
+        } catch (Exception e) {
+            LOGGER.error("An error has occurred obtaining TDB persistentDataset", e);
+        }
+
+        try {
+            temporalDataset = GraphStore.createGraphStore(temporal);
+//            temporalDataset = TDBFactory.createDataset(temporal);
+            if(temporalDataset== null){
+                LOGGER.info("DATASET NULO");
+            }
+        } catch (Exception e) {
+            LOGGER.error("An error has occurred obtaining TDB temporalDataset", e);
+        }
+
+    }
+
+    public GraphStore persistent(){
         return persistentDataset;
     }
+
+    public GraphStore temporal(){
+        return temporalDataset;
+    }
+
+
+//    public Dataset getTDBDataset() {
+//
+//        if (persistentDataset == null) {
+//            try {
+//                persistentDataset = TDBFactory.createDataset(dir+"/"+name);
+//
+//                if(persistentDataset== null){
+//                    LOGGER.info("DATASET NULO");
+//                }
+//            } catch (Exception e) {
+//                LOGGER.error("An error has occurred obtaining TDB persistentDataset");
+//            }
+//        }
+//        return persistentDataset;
+//    }
 
     public void close() {
         if(persistentDataset != null){
@@ -51,6 +122,13 @@ public class JenaConnection {
             persistentDataset.close();
             persistentDataset = null;
         }
+        if(temporalDataset != null){
+            if(temporalDataset.isInTransaction())
+                temporalDataset.commit();
+            temporalDataset.end();
+            temporalDataset.close();
+            temporalDataset = null;
+        }
     }
 
     @PreDestroy
@@ -59,7 +137,7 @@ public class JenaConnection {
         close();
     }
 
-    public static JenaConnection getInstance() {
-        return instance;
-    }
+//    public static JenaConnection getInstance() {
+//        return instance;
+//    }
 }
