@@ -1,8 +1,9 @@
 <template>
   <router-view />
 </template>
+
 <script>
-import { defineComponent } from 'vue';
+import { defineComponent } from 'vue'
 
 export default defineComponent({
   name: 'App'
