@@ -7,19 +7,16 @@ import axios from 'axios'
 // good idea to move this instance creation inside of the
 // "export default () => {}" function below (which runs individually
 // for each client)
-// const api = axios.create({ baseURL: 'https://api.example.com' })
-const odinApi = axios.create({ baseURL: 'http://localhost:8082/odin' });
+const odinApi = axios.create({ baseURL: process.env.API });
 
 export default boot(({ app }) => {
   // for use inside Vue files (Options API) through this.$axios and this.$api
 
-  // app.config.globalProperties.$axios = axios
-  app.config.globalProperties.$axios = axios;
+  app.config.globalProperties.$axios = axios
   // ^ ^ ^ this will allow you to use this.$axios (for Vue Options API form)
   //       so you won't necessarily have to import axios in each vue file
 
-  // app.config.globalProperties.$api = api
-  app.config.globalProperties.$odinApi = odinApi;
+  app.config.globalProperties.$odinApi = odinApi
   // ^ ^ ^ this will allow you to use this.$api (for Vue Options API form)
   //       so you can easily perform requests against your app's API
 })
