@@ -1,5 +1,6 @@
 import { route } from 'quasar/wrappers'
 import { createRouter, createMemoryHistory, createWebHistory, createWebHashHistory } from 'vue-router'
+import { useAuthStore } from 'stores/auth.store.js' 
 import routes from './routes'
 
 /*
@@ -23,7 +24,32 @@ export default route(function (/* { store, ssrContext } */) {
     // Leave this as is and make changes in quasar.conf.js instead!
     // quasar.conf.js -> build -> vueRouterMode
     // quasar.conf.js -> build -> publicPath
-    history: createHistory(process.env.MODE === 'ssr' ? void 0 : process.env.VUE_ROUTER_BASE)
+    history: createHistory(process.env.VUE_ROUTER_BASE)
+  })
+
+
+  Router.beforeEach(async (to, from) => {
+    const authStore = useAuthStore()
+    authStore.init()
+  
+
+      if(!authStore.user.accessToken && to.name !== 'auth') {
+        return { name: 'auth'}
+      }
+
+      // if(authStore.user.accessToken && from.name === 'auth') {
+      //   return {name: 'projects'}
+      // }
+  
+      if(authStore.user.accessToken && to.name === 'auth') {
+        if(from.name === undefined)
+          return {name: 'projects'}
+        else  
+          return false
+      }
+
+ 
+
   })
 
   return Router
